# Preallocate tags and reuse them throughout a request.