--- a/pkg/activator/handler/handler.go
+++ b/pkg/activator/handler/handler.go
@@ -33,7 +33,6 @@ import (
 	activatorconfig "knative.dev/serving/pkg/activator/config"
 	activatornet "knative.dev/serving/pkg/activator/net"
 	"knative.dev/serving/pkg/activator/util"
-	"knative.dev/serving/pkg/apis/serving"
 	pkghttp "knative.dev/serving/pkg/http"
 	"knative.dev/serving/pkg/network"
 	"knative.dev/serving/pkg/queue"
@@ -51,7 +50,6 @@ type Throttler interface {
 type activationHandler struct {
 	transport        http.RoundTripper
 	tracingTransport http.RoundTripper
-	reporter         activator.StatsReporter
 	throttler        Throttler
 	bufferPool       httputil.BufferPool
 }
@@ -60,12 +58,11 @@ type activationHandler struct {
 const defaulTimeout = 2 * time.Minute
 
 // New constructs a new http.Handler that deals with revision activation.
-func New(ctx context.Context, t Throttler, sr activator.StatsReporter) http.Handler {
+func New(ctx context.Context, t Throttler) http.Handler {
 	defaultTransport := pkgnet.AutoTransport
 	return &activationHandler{
 		transport:        defaultTransport,
 		tracingTransport: &ochttp.Transport{Base: defaultTransport},
-		reporter:         sr,
 		throttler:        t,
 		bufferPool:       network.NewBufferPool(),
 	}
@@ -96,12 +93,9 @@ func (a *activationHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 		}, tracingEnabled)
 		proxySpan.End()
 
-		revision := revisionFrom(r.Context())
-		configurationName := revision.Labels[serving.ConfigurationLabelKey]
-		serviceName := revision.Labels[serving.ServiceLabelKey]
 		// Do not report response time here. It is reported in pkg/activator/metric_handler.go to
 		// sum up all time spent on multiple handlers.
-		a.reporter.ReportRequestCount(revID.Namespace, serviceName, configurationName, revID.Name, httpStatus, 1)
+		reporterFrom(r.Context()).ReportRequestCount(httpStatus, 1)
 
 		return nil
 	})