<!--
Request Prow to automatically lint any go code in this PR:

/lint
-->

## Proposed Changes

The activator reports metrics in multiple stages of a request lifecycle. The revision relevant tags will be static throughout the request so this builds a scoped reporter for these tags to reuse them in later stages.

### Before

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	  109980	     10660 ns/op	    4531 B/op	      75 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	  443875	      2716 ns/op	    4491 B/op	      75 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  111962	     11246 ns/op	    4535 B/op	      75 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	  419640	      3512 ns/op	    4519 B/op	      75 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	   98715	     13249 ns/op	    4528 B/op	      75 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	  389748	      3613 ns/op	    4541 B/op	      75 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	   88858	     14147 ns/op	    4537 B/op	      75 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	  426769	      3674 ns/op	    4547 B/op	      75 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	   67862	     16567 ns/op	    4531 B/op	      75 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  329760	      4208 ns/op	    4561 B/op	      75 allocs/op
PASS
```

### After

```
goos: darwin
goarch: amd64
pkg: knative.dev/serving/pkg/activator/handler
BenchmarkHandler/002k-resp-len-sequential-12         	  119520	     10047 ns/op	    4053 B/op	      59 allocs/op
BenchmarkHandler/002k-resp-len-parallel-12           	  485211	      2364 ns/op	    4025 B/op	      59 allocs/op
BenchmarkHandler/016k-resp-len-sequential-12         	  122815	     10546 ns/op	    4057 B/op	      59 allocs/op
BenchmarkHandler/016k-resp-len-parallel-12           	  503700	      2755 ns/op	    4043 B/op	      59 allocs/op
BenchmarkHandler/032k-resp-len-sequential-12         	  116455	     10861 ns/op	    4051 B/op	      59 allocs/op
BenchmarkHandler/032k-resp-len-parallel-12           	  509024	      2626 ns/op	    4050 B/op	      59 allocs/op
BenchmarkHandler/064k-resp-len-sequential-12         	  102384	     14049 ns/op	    4047 B/op	      59 allocs/op
BenchmarkHandler/064k-resp-len-parallel-12           	  462595	      3065 ns/op	    4066 B/op	      59 allocs/op
BenchmarkHandler/128k-resp-len-sequential-12         	   72121	     14986 ns/op	    4049 B/op	      59 allocs/op
BenchmarkHandler/128k-resp-len-parallel-12           	  372339	      3878 ns/op	    4090 B/op	      59 allocs/op
PASS
```

16 allocs saved in the Activator's main handler by reusing the tags.

**Release Note**

<!-- Enter your extended release note in the below block. If the PR requires
additional action from users switching to the new release, include the string
"action required". If no release note is required, write "NONE". -->

```release-note
NONE
```

/assign @vagababov